// packages/admin/generated/graphql.ts
export type Maybe<T> = T | undefined;
export type InputMaybe<T> = T | undefined;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  UInt64: bigint;
};
// ...
export type Asset = Model & {
  __typename?: 'Asset';
  code: Scalars['String'];
  createdAt: Scalars['String'];
  id: Scalars['ID'];
  scale: Scalars['Int'];
  withdrawalThreshold?: Maybe<Scalars['UInt64']>;
};
// ...
export type Http = {
  __typename?: 'Http';
  outgoing: HttpOutgoing;
};

export type HttpIncomingInput = {
  authTokens: Array<Scalars['String']>;
};

export type HttpInput = {
  incoming?: InputMaybe<HttpIncomingInput>;
  outgoing: HttpOutgoingInput;
};

export type HttpOutgoing = {
  __typename?: 'HttpOutgoing';
  authToken: Scalars['String'];
  endpoint: Scalars['String'];
};

export type HttpOutgoingInput = {
  authToken: Scalars['String'];
  endpoint: Scalars['String'];
};
// ...
export type Model = {
  createdAt: Scalars['String'];
  id: Scalars['ID'];
};
// ...
export type MutationResponse = {
  code: Scalars['String'];
  message: Scalars['String'];
  success: Scalars['Boolean'];
};
// ...
export type Peer = Model & {
  __typename?: 'Peer';
  asset: Asset;
  createdAt: Scalars['String'];
  http: Http;
  id: Scalars['ID'];
  maxPacketAmount?: Maybe<Scalars['UInt64']>;
  name?: Maybe<Scalars['String']>;
  staticIlpAddress: Scalars['String'];
};
// ...
export type UpdatePeerInput = {
  http?: InputMaybe<HttpInput>;
  id: Scalars['String'];
  maxPacketAmount?: InputMaybe<Scalars['UInt64']>;
  name?: InputMaybe<Scalars['String']>;
  staticIlpAddress?: InputMaybe<Scalars['String']>;
};

export type UpdatePeerMutationResponse = MutationResponse & {
  __typename?: 'UpdatePeerMutationResponse';
  code: Scalars['String'];
  message: Scalars['String'];
  peer?: Maybe<Peer>;
  success: Scalars['Boolean'];
};